refactor(archival): split the facts validator out, and fix the two style findings
phpmd put ArchivalAnnotationValidator over the complexity threshold and
resolve() over the method-length one, both because of this change.

The three fact validators are now ArchivalFactsValidator. That is a real seam,
not a shuffle: the original validates the disposal DECISION, a duration and
the rules that override it, while these validate FACTS about the record
against MDTO's begrippenlijsten. Moving methods within a class would have
raised complexity rather than lowered it, which is what happened the last time
I tried it.

phpcs also wanted a named argument in the shortened resolver call, and an
inline if replaced by a statement.
